Updated module augmentation for widget package.
