+++ packages/ckeditor5-widget/src/widgettoolbarrepository.ts
@@ -362,9 +362,3 @@ interface WidgetRepositoryToolbarDefinition {
 
 	initialized: boolean;
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ WidgetToolbarRepository.pluginName ]: WidgetToolbarRepository;
-	}
-}
